Split Windows string functions from webview.h
